A website about best places of India.
I'm sure most of us love travelling and loves to visit new places. But sometimes we don't know where we should go, it can be difficult to know about the place, when we don't have any connection at that place. I'm an Indian and I have visited some of the beautiful places. So I have made a website based on India. For the people who wants to visit India. In this I have include all the best places of India. By visiting this site you can know about the place, why is it famous for, how can we reach to it etc. . 
I hope you will like my project helpful.
Thank You